+++ superwall/src/main/java/com/superwall/sdk/paywall/view/webview/SWWebView.kt
@@ -173,7 +173,7 @@ class SWWebView(
         addJavascriptInterface(messageHandler, "SWAndroid")
 
         val webSettings = this.settings
-        setWebContentsDebuggingEnabled(false)
+        setWebContentsDebuggingEnabled(true)
         webSettings.javaScriptEnabled = true
         webSettings.setSupportZoom(false)
         webSettings.builtInZoomControls = false