+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 The changelog for `Superwall`. Also see the [releases](https://github.com/superwall/Superwall-Android/releases) on GitHub.
 
+## 2.6.7
+
+### Enhancements
+- Adds permission granting and callbacks to/from paywalls
+- Adds `PaywallPreloadStart` and `PaywallPreloadComplete` events
+
+### Fixes
+- Fix handling of deep links when paywall is detached
+- Enables permission granting from paywall and callbacks
+- Fix crash when handling drawer style paywalls with 100% height
+ 
 ## 2.6.6
 
 ## Enhancements